frameworks/Kopf/requirements.txt
kubernetes==21.7.0
kopf
pykube-ng[gcp]
pyyaml
